Move scrapeconf struct to scrapeconf.go
