--- pkg/config/runtime_config_test.go
@@ -1,31 +1,0 @@
-// Copyright 2026 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package config
-
-import "testing"
-
-func TestDefaultConfig(t *testing.T) {
-	t.Parallel()
-
-	cfg := DefaultConfig()
-	if cfg.MetricsPerQuery != DefaultMetricsPerQuery {
-		t.Fatalf("MetricsPerQuery = %d, want %d", cfg.MetricsPerQuery, DefaultMetricsPerQuery)
-	}
-	if cfg.TaggingAPIConcurrency != DefaultTaggingAPIConcurrency {
-		t.Fatalf("TaggingAPIConcurrency = %d, want %d", cfg.TaggingAPIConcurrency, DefaultTaggingAPIConcurrency)
-	}
-	if cfg.CloudwatchConcurrency != DefaultCloudwatchConcurrency {
-		t.Fatalf("CloudwatchConcurrency = %+v, want %+v", cfg.CloudwatchConcurrency, DefaultCloudwatchConcurrency)
-	}
-}
